src/requirements.txt
concrete-ml==1.9.0
concrete-python==2.10.0
concrete-ml-extensions==0.1.9
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.5.0
matplotlib==3.7.1
